+++ src/main/java/si/matjazcerkvenik/eventlogger/model/DAlarmSeverity.java
@@ -1,0 +1,13 @@
+package si.matjazcerkvenik.eventlogger.model;
+
+public class DAlarmSeverity {
+
+    public static final int INDETERMINATE = 0;
+    public static final int CRITICAL = 1;
+    public static final int MAJOR = 2;
+    public static final int MINOR = 3;
+    public static final int WARNING = 4;
+    public static final int CLEAR = 5;
+    public static final int INFORMATIONAL = 6;
+
+}